Day10 Finished :panda_face:

+++ Day10/Calculator.py
@@ -28,14 +28,14 @@ from art import logo
 def calculator():
     print(logo)
 
-    num1 = int(input("What's the first number?: "))
+    num1 = float(input("What's the first number?: "))
     for symbol in operations:
         print(symbol)
 
     should_continue = True
     while should_continue:
         operation_symbol = input("Pick an operation from the line above: ")
-        num2 = int(input("What's the second number?: "))
+        num2 = float(input("What's the second number?: "))
         calculation_function = operations[operation_symbol]
         answer = calculation_function(num1, num2)
 
@@ -46,6 +46,6 @@ def calculator():
         else:
             should_continue = False
             calculator()
-        
+
 calculator()
